refactor: split skill-lifecycle.ts into focused modules
Extract frontmatter, graph, prerequisites, register, and evaluate into
separate files under src/core/; keep skill-lifecycle.ts as a barrel.
Drop the abandoned inDegree pre-pass in buildGraph and unify writer
lookups via indexWriters.
